fix: add missing root layout tags
